Upgrade ECharts to v5, add support for ECharts gauge & chart widget (#814)
Add a oh-chart widget to add charts to layout pages; change oh-chart-page to use it.
Add a oh-data-series component with the ability to add gauges representing a single value.
Upgrade ECharts to v5.0.1.

Also-by: Yannick Schaus <github@schaus.net>
