+++ bundles/org.openhab.ui/web/src/assets/definitions/widgets/chart/index.js
@@ -375,6 +375,17 @@ export default {
     }
   },
 
+  'oh-data-series': {
+    label: 'Data Series',
+    docLink: 'https://echarts.apache.org/en/option.html#series',
+    props: {
+      parameterGroups: [],
+      parameters: [
+        seriesTypeParameter('gauge')
+      ]
+    }
+  },
+
   'oh-time-series': {
     label: 'Time Series',
     docLink: 'https://echarts.apache.org/en/option.html#series',
