+++ bundles/org.openhab.ui/web/src/components/widgets/system/index.js
@@ -19,3 +19,4 @@ export { default as OhSwiperSlide } from './oh-swiper-slide.vue'
 export { default as OhTrend } from './oh-trend.vue'
 export { default as OhWebframe } from './oh-webframe.vue'
 export { default as OhRepeater } from './oh-repeater.vue'
+export { default as OhChart } from './oh-chart.vue'
