+++ bundles/org.openhab.ui/web/src/components/widgets/system/oh-chart.vue
@@ -1,0 +1,16 @@
+<template>
+  <oh-chart-component :context="this.context" />
+</template>
+
+<script>
+import mixin from '../widget-mixin'
+import { OhChartDefinition } from '@/assets/definitions/widgets/system'
+
+export default {
+  mixins: [mixin],
+  components: {
+    'oh-chart-component': () => import(/* webpackChunkName: "oh-chart-component" */ './oh-chart-component.vue')
+  },
+  widget: OhChartDefinition
+}
+</script>
